Rename <rox/macros.h> to <rox/defs.h>
